requirements.txt
pytest==8.3.2
fastapi
uvicorn
python-telegram-bot==20.7
httpx
pytest
